Recontruct the code of work function
